Updated to gridpack scripts for strong prod, T2tt, T2bt, TChiWZ and TChiStauStau fragments. Improved validation scripts.

+++ GridpackWorkflow/test/scripts/runFragmentValidation.sh
@@ -1,9 +1,15 @@
 #!/bin/bash
 
+# for local test run inside the directory with the fragment e.g. production/models/T1tttt/
+# the tag can be used when there are multiple datasets per model
+
 JOBNUM=$1
-DATASET=$2
+MODEL=$2
 OUTDIR=$3
 NEVENTS=$4
+LOCAL_TEST=$5
+TAG=$6
+
 
 source /cvmfs/cms.cern.ch/cmsset_default.sh 
 export SCRAM_ARCH=slc6_amd64_gcc530
@@ -16,16 +22,20 @@ cd CMSSW_8_0_5_patch1/src
 eval `scram runtime -sh`
 
 mkdir -p Configuration/GenProduction/python/
-cp ../../fragment.py Configuration/GenProduction/python/
+FRAGMENT=${MODEL}_${TAG}_fragment.py
+OUTPUT=${MODEL}_${TAG}_${JOBNUM}.root
+if [[ -z "${TAG}" ]]; then
+   FRAGMENT=${MODEL}_fragment.py
+   OUTPUT=${MODEL}_${JOBNUM}.root
+fi
+cp ../../${FRAGMENT} Configuration/GenProduction/python/
 
 scram b
 cd ../../
 mkdir test
 cd test
-
-OUTPUT=${DATASET}_${JOBNUM}.root
-
-cmsDriver.py Configuration/GenProduction/python/fragment.py \
+   
+cmsDriver.py Configuration/GenProduction/python/${FRAGMENT} \
 --fileout file:${OUTPUT}  \
 --eventcontent AODSIM \
 --conditions 80X_mcRun2_asymptotic_v12 \
@@ -37,10 +47,11 @@ cmsDriver.py Configuration/GenProduction/python/fragment.py \
 
 cmsRun jobcfg.py
 
-lcg-cp -v -b -D srmv2 --vo cms file:`pwd`/${OUTPUT} srm://bsrm-3.t2.ucsd.edu:8443/srm/v2/server?SFN=${OUTDIR}/${OUTPUT}
-
-cd ../
-rm -rf CMSSW_8_0_5_patch1
-rm -rf test
+if [[ "${LOCAL_TEST}" != 'local' ]]; then 
+   lcg-cp -v -b -D srmv2 --vo cms file:`pwd`/${OUTPUT} srm://bsrm-3.t2.ucsd.edu:8443/srm/v2/server?SFN=${OUTDIR}/${OUTPUT}    
+   cd ../
+   rm -rf CMSSW_8_0_5_patch1
+   rm -rf test
+fi
 
 echo "Bye."
